Fix step1 test on newer gcc/gdb versions
On one of our machines this test is failing. Upon investigation,
the reason for this is that the breakpoint for a function is
considered by gdb to be on the line of the opening brace, while
the test driver for this test assumes that the breakpoint for the
function is on the line of the first statement. This is likely
a change in the debug information emitted by the compiler on this
machine (gcc (Ubuntu 9.2.1-31ubuntu1) 9.2.1 20200304), but may also
be a change in GDB (GNU gdb (Ubuntu 9.0.50.20191119-0ubuntu1) 9.0.50.20191119-git).
In either case, it's easy to paper over the difference by making sure
that the opening brace and the first statement are on the same line
(even if it looks a bit ugly).

(cherry picked from commit 5ffa8be481cf5ae4b01c922c5c4ad4c69d7bea2b)