string_instructions_multiwatch: Work around buggy compiler DWARF
Using the same gcc/gdb versions as in #2457, I see the
string_instructions_multiwatch test fail, because gdb fails to
properly print the location of the `src` variable in my_memmove.
The reason for this is that it has not yet executed the part
of the prologue that moves the src value from its ABI register into
the register assigned to it for the rest of the function as recorded
in DWARF. However, it didn't mark the instructions that do that as
part of the prologue either. To me, this is a compiler bug, but
given the poor state of DWARF emission in mainstream compilers,
I can't say I'm particularly surprised. This fixes the test by
having the test itself print out the location of the `buf` variable,
rather than relying on the DWARF to be accurate.

(cherry picked from commit ce1dfc3ee098770be2c4b7c91db0e081210ca34e)