Implement send_fd
This function is the opposite of retrieve_fd. Rather than transferring
an fd from the tracee to us, it injects one of our fds into the tracee's
fd table. While we're here, refactor the entire thing to do less manual
manipulation of data structure offsets. The C++ compiler is quite good at
that, so let it take care of it.

(cherry picked from commit 0ec1068181b63b11817035329f2bacdf57a01008)