Get rid of RR_RESERVED_ROOT_DIR_FD
Admittedly, this change is primarily motivated by a philisophical
objection to RR_RESERVED_ROOT_DIR_FD, rather than any great practical
need. It just seems to me that if the tracee went through all the
trouble of hiding the real root, then we shouldn't just give it an
easy way to get it back. Of course, I'm not suggesting that this
would be the only issue with running an external rr tracer on a
sanboxed process without compromising that sandbox's security
guarantees - I'm sure there's many others. However, this one's
a pretty obvious one, so I'd like to remove it if possible.

That said, there are a couple of minor practical benefits:
- It's more robust to weird setups (weird mount setups, tmp being
  a symlink).
- It's more robust when recording programs that change uid
- As a result, we can get rid of a few fallbacks that would only
  get used if we happened to be in one of these weird setups (e.g.
  where we would fail to map the rr page).
- We can use memfd in more cases. This is a slightly benefit if /tmp
  happens to not be a tmpfs, which happens occasionally.
- There was a note in the code that things could fail in the nesting
  scenario if somebody chroot'ed between the outer and the inner rr
  (e.g. by recording a container that itself launched an rr process).
  That should work fine now.
- We avoid a corner case where after a setuid, some files in the trace
  directory would have the wrong uid.

None of these are particularly important, but together, I find
them convincing enough that I think this is worth doing (I happened
to need the send_fd functionality this depends upon for an
unrelated change, which made doing this quite easy).

The one place where this is slightly annoying is in open_mem_fd, because
we now need to send an fd and then get a new fd back. However, that
only needs to happen if rr fails to open /proc/<pid>/mem. A comment
in the code says that Ubuntu kernels have this behavior. However,
I was unable to reproduce on current Ubuntu kernels. It's also
counterbalanced by allowing us to do slightly fewer syscalls elsewhere.
And lastly, with pidfd_getfd and related functionality, these kinds
of file transfers might in the future be quite a bit faster anyway.

(cherry picked from commit 25c93ba12c2e05d355eb599727bf9c39c95e99a1)