Make failing to dup the desched fd non-fatal
We try to avoid using the low fds to avoid disrupting the tracee (e.g.
the tracee may want to reserve a portion of fd space for its exclusive
management and get confused if we put an fd there). However, if we can't
use a high fd (e.g. because of a ulimit), then this shouldn't be fatal,
since it's not a problem if the tracee doesn't indiscriminately close
file descriptors. We could probably do better in the future, but for
now, let's just close the immediate failure mode with low ulimits.

Part of #2521

(cherry picked from commit 7254a89d1ab7919c8d92d26a19e0a24afd4f4cf1)