Don't patch sigreturn syscalls.
In kernel 5.8 the layout of the 32 bit VDSO has changed. The sigreturn piece of the VDSO is now followed by several nops which allows our syscall monkeypatcher to work. That's really bad though, because syscallbuf will switch the stack and when the sigreturn syscall hits the kernel the sigframe is not present on the stack where it should be.

Fixes #2635.

(cherry picked from commit fcd2a26680a3fc2bda5f40d732d0c72b9628357b)